Added more things to x86_64 image.
Some preference apps, mount_server and AboutSystem. Removed the check
for x86_64 in the boot script, the normal path through the script will
work now. Also removed a temporary hack to workaround AboutSystem not
being there in build_haiku_image.

+++ src/preferences/time/NetworkTimeView.cpp
@@ -479,7 +479,7 @@ NetworkTimeView::_InitView()
 void
 NetworkTimeView::_UpdateServerList()
 {
-	while (fServerListView->RemoveItem(0L) != NULL);
+	while (fServerListView->RemoveItem((int32)0) != NULL);
 
 	const char* server;
 	int32 index = 0;
